rename paths after merge
