+++ internal/generate/gen/config.go
